+++ Sources/Datadog/RUM/RUMTelemetry.swift
@@ -29,7 +29,7 @@ internal final class RUMTelemetry: Telemetry {
     var configurationEventMapper: RUMTelemetryConfiguratoinMapper?
     let delayedDispatcher: RUMTelemetryDelayedDispatcher
     let sampler: Sampler
-    var configurationExtraSampler = Sampler(samplingRate: 0.2)
+    var configurationExtraSampler = Sampler(samplingRate: 20.0)
 
     /// Keeps track of current session
     @ReadWriteLock
@@ -51,7 +51,7 @@ internal final class RUMTelemetry: Telemetry {
         configurationEventMapper: RUMTelemetryConfiguratoinMapper?,
         delayedDispatcher: RUMTelemetryDelayedDispatcher?,
         sampler: Sampler,
-        configurationExtraSampler: Sampler = Sampler(samplingRate: 0.2) // Extra sample of 20% for configuration events
+        configurationExtraSampler: Sampler = Sampler(samplingRate: 20.0) // Extra sample of 20% for configuration events
     ) {
         self.core = core
         self.dateProvider = dateProvider
@@ -236,7 +236,10 @@ private extension FeaturesConfiguration {
             initializationType: nil,
             mobileVitalsUpdatePeriod: self.rum?.vitalsFrequency?.toInt64Milliseconds,
             premiumSampleRate: nil,
+            reactNativeVersion: nil,
+            reactVersion: nil,
             replaySampleRate: nil,
+            selectedTracingPropagators: nil,
             sessionReplaySampleRate: nil,
             sessionSampleRate: self.rum?.sessionSampler.samplingRate.toInt64(),
             silentMultipleInit: nil,
@@ -258,6 +261,7 @@ private extension FeaturesConfiguration {
             trackSessionAcrossSubdomains: nil,
             trackViewsManually: nil,
             useAllowedTracingOrigins: nil,
+            useAllowedTracingUrls: nil,
             useBeforeSend: nil,
             useCrossSiteSessionCookie: nil,
             useExcludedActivityUrls: nil,