+++ Tests/DatadogTests/DatadogObjc/DDRUMMonitorTests.swift
@@ -175,7 +175,9 @@ class DDRUMMonitorTests: XCTestCase {
 
         let rumEventMatchers = try core.waitAndReturnRUMEventMatchers()
 
-        let viewEvents = rumEventMatchers.filterRUMEvents(ofType: RUMViewEvent.self)
+        let viewEvents = rumEventMatchers.filterRUMEvents(ofType: RUMViewEvent.self) { event in
+            return event.view.name != RUMOffViewEventsHandlingRule.Constants.applicationLaunchViewName
+        }
         XCTAssertEqual(viewEvents.count, 4)
 
         let event1: RUMViewEvent = try viewEvents[0].model()
@@ -208,7 +210,9 @@ class DDRUMMonitorTests: XCTestCase {
 
         let rumEventMatchers = try core.waitAndReturnRUMEventMatchers()
 
-        let viewEvents = rumEventMatchers.filterRUMEvents(ofType: RUMViewEvent.self)
+        let viewEvents = rumEventMatchers.filterRUMEvents(ofType: RUMViewEvent.self) { event in
+            return event.view.name != RUMOffViewEventsHandlingRule.Constants.applicationLaunchViewName
+        }
         XCTAssertEqual(viewEvents.count, 3)
 
         let event1: RUMViewEvent = try viewEvents[0].model()
@@ -392,7 +396,9 @@ class DDRUMMonitorTests: XCTestCase {
 
         let rumEventMatchers = try core.waitAndReturnRUMEventMatchers()
 
-        let viewEvents = rumEventMatchers.filterRUMEvents(ofType: RUMViewEvent.self)
+        let viewEvents = rumEventMatchers.filterRUMEvents(ofType: RUMViewEvent.self) { event in
+            return event.view.name != RUMOffViewEventsHandlingRule.Constants.applicationLaunchViewName
+        }
         XCTAssertEqual(viewEvents.count, 1)
 
         XCTAssertEqual(try viewEvents[0].attribute(forKeyPath: "context.global-attribute1"), "foo1")